% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjroc.R
\name{adjroc}
\alias{adjroc}
\title{adjroc}
\usage{
adjroc(
  score,
  class,
  method = "emp",
  sensitivity = NULL,
  specificity = NULL,
  plot = FALSE
)
}
\arguments{
\item{score}{An numeric array of diagnostic score}

\item{class}{An array of equal length of \code{"score"}, containing the class of the observations}

\item{method}{Specifies the method for estimating the ROC curve. Three methods are supported, which are \code{"empirical"}, \code{"binormal"}, and \code{"nonparametric"}}

\item{sensitivity}{numeric. Specify the threshold of sensitivity}

\item{specificity}{numeric. Specify the threshold of specificity}

\item{plot}{logical. if TRUE, the sensitivity and specificity will be plotted}
}
\description{
computes adjusted sensitivity, adjusted specificity, or the meeting
    point between sensitivity and specificity
}
\examples{
\dontrun{
# from ROCit R package
library(ROCit)
data("Diabetes")

# calculate adjusted sensitivity, when specificity threshold is 0.90:
sen <- adjroc(score = Diabetes$chol, class = Diabetes$dtest)
roc_binormal <- rocit(score = Diabetes$chol, class = Diabetes$dtest,
 negref = "-", method = "bin")
geocode(dat)
}
}
